Crate re_data_store

The Rerun datastore, implemented on top of Apache Arrow using the [arrow2] crate.

This crate is an in-memory time series database for Rerun log data. It is indexed by Entity path, component, timeline, and time. It supports out-of-order insertions, and fast O(log(N)) queries.

See DataStore for an overview of the core data structures.

See DataStore::latest_at and DataStore::range for the documentation of the public read APIs.

See DataStore::insert_row for the documentation of the public write APIs.

Feature flags

deadlock_detection — Enables parking_lot’s deadlock detection background thread.

Re-exports

pub use arrow2::io::ipc::read::StreamReader;

pub use arrow2::io::ipc::read::StreamState;

pub use re_log_types::ResolvedTimeRange;

pub use re_log_types::TimeInt;

pub use re_log_types::TimeType;

pub use re_log_types::Timeline;

Modules

arrow_util 🔒

external

store 🔒

store_arrow 🔒

store_dump 🔒

store_event 🔒

store_format 🔒

store_gc 🔒

store_helpers 🔒

store_read 🔒

store_sanity 🔒

store_stats 🔒

store_subscriber 🔒

store_write 🔒

Structs

DataStore

A complete data store: covers all timelines, all entities, everything.

DataStoreConfig

DataStoreRowStats

DataStoreStats

EntityStats

GarbageCollectionOptions

LatestAtQuery

A query at a given time, for a given timeline.

RangeQuery

A query over a time range, for a given timeline.

StoreDiff

Describes an atomic change in the Rerun DataStore: a row has been added or deleted.

StoreEvent

The atomic unit of change in the Rerun DataStore.

StoreGeneration

Incremented on each edit.

StoreSubscriberHandle

Enums

GarbageCollectionTarget

StoreDiffKind

Is it an addition or a deletion?

WriteError

Traits

ArrayExt

StoreSubscriber

A StoreSubscriber subscribes to atomic changes from all DataStores through StoreEvents.

Type Aliases

WriteResult