Struct re_data_store::store::IndexedTable

pub struct IndexedTable {
    pub timeline: Timeline,
    pub entity_path: EntityPath,
    pub buckets: BTreeMap<TimeInt, IndexedBucket>,
    pub all_components: ComponentNameSet,
    pub buckets_num_rows: u64,
    pub buckets_size_bytes: u64,
}

An IndexedTable is an ever-growing, arbitrary large [re_log_types::DataTable] that is optimized for time-based insertions and queries (which means a lot of bucketing).

See also IndexedBucket.

Run the following command to display a visualization of the store’s internal datastructures and better understand how everything fits together:

cargo test -p re_data_store -- --nocapture datastore_internal_repr

Fields
timeline: Timeline

The timeline this table operates in, for debugging purposes.

entity_path: EntityPath

The entity this table is related to, for debugging purposes.

buckets: BTreeMap<TimeInt, IndexedBucket>

The actual buckets, where the data is stored.

The keys of this BTreeMap represent the lower bounds of the time-ranges covered by their associated buckets, as seen from an indexing rather than a data standpoint!

This means that e.g. for the initial bucket, this will always be -∞, as from an indexing standpoint, all reads and writes with a time t >= -∞ should go there, even though the bucket doesn’t actually contains data with a timestamp of -∞!

all_components: ComponentNameSet

Track all of the components that have been written to.

Note that this set will never be purged and will continue to return components that may have been set in the past even if all instances of that component have since been purged to free up space.

buckets_num_rows: u64

The number of rows stored in this table, across all of its buckets.

buckets_size_bytes: u64

The size of both the control & component data stored in this table, across all of its buckets, in bytes.

This is a best-effort approximation, adequate for most purposes (stats, triggering GCs, …).

Implementations

impl IndexedTable

pub fn new(timeline: Timeline, entity_path: EntityPath) -> Self

pub(crate) fn uphold_indexing_invariants(&mut self)

Makes sure bucketing invariants are upheld, and takes necessary actions if not.

Invariants are:

There must always be at least one bucket alive.

The first bucket must always have an indexing time -∞.

impl IndexedTable

fn try_drop_bucket(&mut self, max_row_id: RowId) -> (Vec<StoreDiff>, u64)

Try to drop an entire bucket at once if it doesn’t contain any RowId greater than max_row_id.

fn try_drop_row( &mut self, row_id: RowId, time: TimeInt ) -> (Option<StoreDiff>, u64)

Tries to drop the given row_id from the table, which is expected to be found at the specified time.

Returns how many bytes were actually dropped, or zero if the row wasn’t found.

impl IndexedTable

pub fn latest_at<const N: usize>( &self, query_time: TimeInt, primary: ComponentName, component_names: &[Option<ComponentName>; N] ) -> Option<(TimeInt, RowId, [Option<DataCell>; N])>

Queries the table for the cells of the specified component_names, as seen from the point of view of the so-called primary component.

Returns an array of [DataCell]s (as well as the associated data time and RowId) on success, or None iff no cell could be found for the primary component.

pub fn range<const N: usize>( &self, time_range: ResolvedTimeRange, component_names: [Option<ComponentName>; N] ) -> impl Iterator<Item = (TimeInt, RowId, [Option<DataCell>; N])> + '_

Iterates the table in order to return the cells of the specified component_names for the given time range.

For each and every relevant row that is found, the returned iterator will yield an array that is filled with the cells of each and every component in component_names, or None if said component is not available in that row.

This method cannot fail! If there’s no data to return, an empty iterator is returned.

pub fn find_bucket(&self, time: TimeInt) -> (TimeInt, &IndexedBucket)

Returns the indexed bucket whose time range covers the given time.

In addition to returning a reference to the IndexedBucket itself, this also returns its indexing time, which is different from its minimum time range bound!

See IndexedTable::buckets for more information.

pub fn find_bucket_mut( &mut self, time: TimeInt ) -> (TimeInt, &mut IndexedBucket)

Returns the indexed bucket whose time range covers the given time.

In addition to returning a reference to the IndexedBucket itself, this also returns its indexing time, which is different from its minimum time range bound!

See IndexedTable::buckets for more information.

pub fn range_buckets( &self, time_range: impl RangeBounds<TimeInt> ) -> impl Iterator<Item = (TimeInt, &IndexedBucket)>

Returns an iterator that is guaranteed to yield at least one bucket, which is the bucket whose time range covers the start bound of the given time_range.

It then continues yielding buckets until it runs out, in increasing time range order.

In addition to yielding references to the IndexedBuckets themselves, this also returns their indexing times, which are different from their minimum time range bounds!

See IndexedTable::buckets for more information.

pub fn range_buckets_rev( &self, time_range: impl RangeBounds<TimeInt> ) -> impl Iterator<Item = (TimeInt, &IndexedBucket)>

Returns an iterator that is guaranteed to yield at least one bucket, which is the bucket whose time range covers the end bound of the given time_range.

It then continues yielding buckets until it runs out, in decreasing time range order.

In addition to yielding references to the IndexedBuckets themselves, this also returns their indexing times, which are different from their minimum time range bounds!

See IndexedTable::buckets for more information.

pub fn range_bucket_rev_mut( &mut self, time_range: impl RangeBounds<TimeInt> ) -> impl Iterator<Item = (TimeInt, &mut IndexedBucket)>

Returns an iterator that is guaranteed to yield at least one bucket, which is the bucket whose time range covers the end bound of the given time_range.

It then continues yielding buckets until it runs out, in decreasing time range order.

In addition to yielding references to the IndexedBuckets themselves, this also returns their indexing times, which are different from their minimum time range bounds!

See IndexedTable::buckets for more information.

pub fn sort_indices_if_needed(&self)

Sort all unsorted indexed buckets in this table.

impl IndexedTable

pub fn sanity_check(&self) -> Result<(), SanityError>

Runs the sanity check suite for the entire table.

Returns an error if anything looks wrong.

impl IndexedTable

pub fn num_rows(&self) -> u64

Returns the number of rows stored across this entire table, i.e. the sum of the number of rows stored across all of its buckets.

pub(crate) fn num_rows_uncached(&self) -> u64

Returns the number of rows stored across this entire table, i.e. the sum of the number of rows stored across all of its buckets.

Recomputed from scratch, for sanity checking.

pub(crate) fn size_bytes_uncached(&self) -> u64

pub fn num_buckets(&self) -> u64

Returns the number of buckets stored across this entire table.

pub fn time_range(&self) -> ResolvedTimeRange

The time range covered by this table.

impl IndexedTable

pub fn insert_row( &mut self, config: &DataStoreConfig, insert_id: Option<u64>, time: TimeInt, row: &DataRow )