Module re_data_store::store_write

Enums

WriteError

Functions

find_split_index 🔒

Finds an optimal split point for the given time index, or None if all entries in the index are identical, making it unsplittable.

split_time_range_off 🔒

Given a time index and a desired split index, splits off the given time range in place, and returns a new time range corresponding to the second part.

Type Aliases

WriteResult